+++ commands/key_rotate_submission_test.go
@@ -2,7 +2,6 @@ package commands_test
 
 import (
 	"testing"
-	"time"
 
 	"code.vegaprotocol.io/protos/commands"
 	commandspb "code.vegaprotocol.io/protos/vega/commands/v1"
@@ -19,48 +18,58 @@ func TestSubmittingNoKeyRotateSubmissionCommandFails(t *testing.T) {
 func TestKeyRotateSubmissionSubmittingEmptyCommandFails(t *testing.T) {
 	err := checkKeyRotateSubmission(&commandspb.KeyRotateSubmission{})
 
-	assert.Contains(t, err.Get("key_rotate_submission.new_pub_key_hash"), commands.ErrIsRequired)
-	assert.Contains(t, err.Get("key_rotate_submission.key_number"), commands.ErrIsRequired)
-	assert.Contains(t, err.Get("key_rotate_submission.time"), commands.ErrNotAValidInteger)
+	assert.Contains(t, err.Get("key_rotate_submission.new_pub_key"), commands.ErrIsRequired)
+	assert.Contains(t, err.Get("key_rotate_submission.new_pub_key_index"), commands.ErrIsRequired)
+	assert.Contains(t, err.Get("key_rotate_submission.current_pub_key_hash"), commands.ErrIsRequired)
+	assert.Contains(t, err.Get("key_rotate_submission.target_block"), commands.ErrIsRequired)
 }
 
-func TestKeyRotateSubmissionMissingNewPubKeyHashFails(t *testing.T) {
+func TestKeyRotateSubmissionMissingNewPubKeyFails(t *testing.T) {
 	err := checkKeyRotateSubmission(&commandspb.KeyRotateSubmission{
-		KeyNumber:   10,
-		TargetBlock: 100,
-		Time:        time.Now().UnixNano(),
+		NewPubKeyIndex:    10,
+		TargetBlock:       100,
+		CurrentPubKeyHash: "w3werertdg",
 	})
 
-	assert.Contains(t, err.Get("key_rotate_submission.new_pub_key_hash"), commands.ErrIsRequired)
+	assert.Contains(t, err.Get("key_rotate_submission.new_pub_key"), commands.ErrIsRequired)
 }
 
-func TestKeyRotateSubmissionMissingKeyNumberFails(t *testing.T) {
+func TestKeyRotateSubmissionMissingNewPubKeyIndexFails(t *testing.T) {
 	err := checkKeyRotateSubmission(&commandspb.KeyRotateSubmission{
-		NewPubKeyHash: "123456789abcdef",
-		TargetBlock:   100,
-		Time:          time.Now().UnixNano(),
+		NewPubKey:         "123456789abcdef",
+		TargetBlock:       100,
+		CurrentPubKeyHash: "w3werertdg",
 	})
 
-	assert.Contains(t, err.Get("key_rotate_submission.key_number"), commands.ErrIsRequired)
+	assert.Contains(t, err.Get("key_rotate_submission.new_pub_key_index"), commands.ErrIsRequired)
 }
 
-func TestKeyRotateSubmissionMissingTimeFails(t *testing.T) {
+func TestKeyRotateSubmissionMissingCurrentPubKeyHashFails(t *testing.T) {
 	err := checkKeyRotateSubmission(&commandspb.KeyRotateSubmission{
-		KeyNumber:     10,
-		NewPubKeyHash: "123456789abcdef",
-		TargetBlock:   100,
-		Time:          -1,
+		NewPubKey:      "123456789abcdef",
+		NewPubKeyIndex: 10,
+		TargetBlock:    100,
 	})
 
-	assert.Contains(t, err.Get("key_rotate_submission.time"), commands.ErrNotAValidInteger)
+	assert.Contains(t, err.Get("key_rotate_submission.current_pub_key_hash"), commands.ErrIsRequired)
+}
+
+func TestKeyRotateSubmissionMissingTargetBlockFails(t *testing.T) {
+	err := checkKeyRotateSubmission(&commandspb.KeyRotateSubmission{
+		NewPubKey:         "123456789abcdef",
+		NewPubKeyIndex:    10,
+		CurrentPubKeyHash: "w3werertdg",
+	})
+
+	assert.Contains(t, err.Get("key_rotate_submission.target_block"), commands.ErrIsRequired)
 }
 
 func TestSubmittingEmptyCommandSuccess(t *testing.T) {
 	err := checkKeyRotateSubmission(&commandspb.KeyRotateSubmission{
-		KeyNumber:     10,
-		NewPubKeyHash: "123456789abcdef",
-		TargetBlock:   100,
-		Time:          time.Now().UnixNano(),
+		NewPubKeyIndex:    10,
+		NewPubKey:         "123456789abcdef",
+		TargetBlock:       100,
+		CurrentPubKeyHash: "w3werertdg",
 	})
 
 	assert.True(t, err.Empty())