+++ vega/snapshot/v1/snapshot.validator.pb.go
@@ -1021,11 +1021,21 @@ func (this *StakeVerifierRemoved) Validate() error {
 func (this *StakeVerifierPending) Validate() error {
 	return nil
 }
+func (this *PendingKeyRotation) Validate() error {
+	return nil
+}
 func (this *Topology) Validate() error {
 	for _, item := range this.ValidatorData {
 		if item != nil {
 			if err := github_com_mwitkow_go_proto_validators.CallValidatorIfExists(item); err != nil {
 				return github_com_mwitkow_go_proto_validators.FieldError("ValidatorData", err)
+			}
+		}
+	}
+	for _, item := range this.PendingPubKeyRotations {
+		if item != nil {
+			if err := github_com_mwitkow_go_proto_validators.CallValidatorIfExists(item); err != nil {
+				return github_com_mwitkow_go_proto_validators.FieldError("PendingPubKeyRotations", err)
 			}
 		}
 	}